fix: move addon to the separate entry point
prevent importing browser-specific code to the node
