+++ src/client/addon/index.ts
@@ -1,0 +1,2 @@
+export * from './withCreevey';
+export * from './readyForCapture';
