+++ src/index.ts
@@ -1,5 +1,3 @@
 export * from './types';
-export * from './client/addon/withCreevey';
-export * from './client/addon/readyForCapture';
 export { loadStories as browserStoriesProvider } from './server/storybook/providers/browser';
 export { loadStories as nodejsStoriesProvider } from './server/storybook/providers/nodejs';
