+++ stories/ImagesViews.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas } from '@storybook/addon-docs';
 import { fireEvent, within } from '@storybook/testing-library';
 
-import { capture } from '../src/client/addon/withCreevey';
+import { capture } from '../src/client/addon';
 import { ImagesView as ImagesViewBase } from '../src/client/shared/components/ImagesView';
 import { ImagesViewMode, CSFStory } from '../src/types';
